chore: move 'ProductForm' into 'dto' package
